requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.8
django-bootstrap4==3.0.1
django-extra-views==0.14.0
django-mathfilters==1.0.0
djangorestframework==3.12.4
pytz==2021.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.2
